Mouse mode (with persistent toggle)
The React-DnD library has two backends, HTML5 (mouse) and Touch. While the Touch backend can be configured to _also_ react to mouse events, the results are not ideal (for instance, dragging the mouse pointer also starts a text selection).

Thus, for the moment, the user will have to toggle between the two modes on each browser/device. Touch mode is the default, because it at least _somewhat_ supports the mouse.