Persist character
Rewrite `App.tsx` to maintain 1 character object as its state. Hook into the `set` function to persist the character as a side effect.

_Technically_ the character object/aggregate should be immutable to be used as state, but immutable objects are a pain in TS/JS. We instead rely on deep cloning (via lodash.deepclone) to ensure that we don't mutate React state in-place.

Data is stored in local storage. This is fine, _except_ when using the app across multiple tabs. Currently, the last write wins. Maybe we can detect when we get focus and re-load from storage?