Implement undo/redo
We store a list of undo/redo actions (currently limited to 10 actions). The actions are persisted to local storage (separately from characters).

Every modification to characters takes a detour through an "UndoManager". That class

1. turns the modification into an action (we call this "forward action")
2. computes the "opposite" action (the "undo action")
3. performs the forward action
4. commits the undo action

This is very close to a WAL (except that we don't _write_ before we perform the action).

The undo/redo actions are stored separately from character data so that I can safely nuke the undo/redo data and keep the current character state. The undo/redo actions might experience more breaking changes than the characters.